# Conflicts:
#	.gitlab-ci-seo.yml
#	package-lock.json
#	package.json
#	pages/+Head.vue
#	pages/metadata/+Page.vue
#	pages/metadata/@id/+Page.vue

+++ src/modules/user/components/DashboardPage.vue
@@ -9,10 +9,7 @@
       color="accent"
     />
     <div v-if="!user" class="notSignedinGrid">
-      <NotFoundCard
-        v-bind="notSignedInInfos"
-        :actionButtonCallback="catchSigninClick"
-      />
+      <NotFoundCard v-bind="notSignedInInfos" :actionButtonCallback="catchSigninClick" />
     </div>
     <div v-if="user" class="dashboardGrid">
       <div class="topBoard mt-6 mt-md-0">
@@ -21,14 +18,10 @@
           :introText="userDashboardConfig.introText"
           :feedbackText="userDashboardConfig.feedbackText"
           :oldDashboardUrl="oldDashboardUrl"
-          :createClickCallback="
-            organizationsStore.canCreateDatasets ? createClickCallback : null
-          "
+          :createClickCallback="organizationsStore.canCreateDatasets ? createClickCallback : null"
           :editingClickCallback="editingClickCallback"
-          :editingDatasetName="lastEditedDataset"
-          :currentLocalDataset="
-            hasLocalDataset ? currentLocalDataset : undefined
-          "
+          :editingDatasetName="lastDataset"
+          :currentLocalDataset="hasLocalDataset ? currentLocalDataset : undefined"
           @localCardClicked="catchLocalCardClick"
           @clearButtonClicked="catchClearLocalStorage"
         />
@@ -45,12 +38,7 @@
           @organizationClick="catchOrganizationClick"
         />
 
-        <FlipLayout
-          v-if="userEditingEnabled"
-          :height="userCardHeight"
-          :width="userCardWidth"
-          :autoButtonFlip="true"
-        >
+        <FlipLayout v-if="userEditingEnabled" :height="userCardHeight" :width="userCardWidth" :autoButtonFlip="true">
           <template v-slot:front>
             <UserCard
               :height="userCardHeight"
@@ -98,18 +86,9 @@
           :clickCallback="catchRefreshClick"
         />
 
-        <div
-          v-if="hasUserDatasets && userDatasetsLoading"
-          id="collaboratorPlaceholders"
-          class="datasetsOverflow"
-        >
+        <div v-if="hasUserDatasets && userDatasetsLoading" id="collaboratorPlaceholders" class="datasetsOverflow">
           <v-row no-gutters>
-            <v-col
-              v-for="n in placeHolderAmount"
-              :key="`userDatasetsPlaceholder_${n}`"
-              cols="2"
-              class="pa-2"
-            >
+            <v-col v-for="n in placeHolderAmount" :key="`userDatasetsPlaceholder_${n}`" cols="2" class="pa-2">
               <MetadataCardPlaceholder />
             </v-col>
           </v-row>
@@ -142,31 +121,19 @@
         <div v-if="!hasUserDatasets" class="noUserDatasetsGrid px-1">
           <NotFoundCard
             v-bind="noDatasetsInfos"
-            :actionButtonCallback="
-              isEditorAndAbove ? createClickCallback : null
-            "
+            :actionButtonCallback="isEditorAndAbove ? createClickCallback : null"
           />
 
-          <NotificationCard
-            v-if="noUserDatasetsError"
-            v-bind="noUserDatasetsError"
-            :showCloseButton="false"
-          />
+          <NotificationCard v-if="noUserDatasetsError" v-bind="noUserDatasetsError" :showCloseButton="false" />
         </div>
       </div>
 
-      <div
-        class="midBoard pt-4"
-        id="collaboratorDatasets"
-        ref="collaboratorDatasets"
-      >
+      <div class="midBoard pt-4" id="collaboratorDatasets" ref="collaboratorDatasets">
         <TitleCard
           title="My Collaborator Datasets"
           :icon="mdiRefresh"
           :tooltipText="refreshButtonText"
-          :loading="
-            collaboratorDatasetIdsLoading || collaboratorDatasetsLoading
-          "
+          :loading="collaboratorDatasetIdsLoading || collaboratorDatasetsLoading"
           :clickCallback="catchCollaboratorRefreshClick"
         />
 
@@ -176,23 +143,14 @@
           class="datasetsOverflow"
         >
           <v-row no-gutters>
-            <v-col
-              v-for="n in orgaDatasetsPreview"
-              :key="`orgaDatasetsPlaceholder_${n}`"
-              cols="2"
-              class="pa-2"
-            >
+            <v-col v-for="n in orgaDatasetsPreview" :key="`orgaDatasetsPlaceholder_${n}`" cols="2" class="pa-2">
               <MetadataCardPlaceholder id="orgaDataset" />
             </v-col>
           </v-row>
         </div>
 
         <div
-          v-if="
-            !collaboratorDatasetIdsLoading &&
-            !collaboratorDatasetsLoading &&
-            hasCollaboratorDatasets
-          "
+          v-if="!collaboratorDatasetIdsLoading && !collaboratorDatasetsLoading && hasCollaboratorDatasets"
           id="collaboratorDatasetCards"
           class="datasetsOverflow"
         >
@@ -229,11 +187,7 @@
         </div>
 
         <div
-          v-if="
-            !collaboratorDatasetIdsLoading &&
-            !collaboratorDatasetsLoading &&
-            !hasCollaboratorDatasets
-          "
+          v-if="!collaboratorDatasetIdsLoading && !collaboratorDatasetsLoading && !hasCollaboratorDatasets"
           class="noUserDatasetsGrid px-1"
         >
           <NotFoundCard v-bind="noCollaboratorDatasetsInfos" />
@@ -251,7 +205,7 @@
  * @author Dominik Haas-Artho
  *
  * Created at     : 2020-07-14 14:18:32
- * Last modified  : 2020-10-13 12:49:34
+ * Last modified  : 2025-10-02 13:16:20
  */
 
 import { mapState, mapGetters } from 'vuex';
@@ -281,29 +235,18 @@ import {
   USER_SIGNIN_PATH,
   METADATADETAIL_PAGENAME,
   METADATAEDIT_PAGENAME,
-  METADATA_CREATION_PATH,
   METADATA_CREATION_PAGENAME,
   ORGANIZATIONS_PAGENAME,
+  WORKFLOW_PAGENAME,
 } from '@/router/routeConsts';
 
 import { useOrganizationsStore } from '@/modules/organizations/store/organizationsStorePinia';
 import { getNameInitials } from '@/factories/authorFactory';
 import { errorMessage } from '@/factories/notificationFactory';
-import {
-  getTagColor,
-  getPopularTags,
-  tagsIncludedInSelectedTags,
-} from '@/factories/keywordsFactory';
-import {
-  enhanceMetadataEntry,
-  isTagSelected,
-} from '@/factories/metaDataFactory';
+import { getTagColor, getPopularTags, tagsIncludedInSelectedTags } from '@/factories/keywordsFactory';
+import { enhanceMetadataEntry, isTagSelected } from '@/factories/metaDataFactory';
 
-import {
-  getUserOrganizationRoleMap,
-  hasOrganizationRoles,
-  isMember,
-} from '@/factories/userEditingValidations';
+import { getUserOrganizationRoleMap, hasOrganizationRoles, isMember } from '@/factories/userEditingValidations';
 
 import UserNotFound1 from '@/modules/user/assets/UserNotFound1.jpg';
 import UserNotFound2 from '@/modules/user/assets/UserNotFound2.jpg';
@@ -319,8 +262,6 @@ import {
 
 import { getPreviewDatasetFromLocalStorage } from '@/factories/userCreationFactory';
 
-import { METADATA_TITLE_PROPERTY } from '@/factories/metadataConsts';
-
 import { loadRouteTags } from '@/factories/stringFactory';
 import categoryCards from '@/store/categoryCards';
 
@@ -333,19 +274,14 @@ import EditUserProfile from '@/modules/user/components/edit/EditUserProfile.vue'
 import FlipLayout from '@/components/Layouts/FlipLayout.vue';
 
 import { getMetadataVisibilityState } from '@/factories/publicationFactory';
+import { useDatasetWorkflowStore } from '@/modules/workflow/datasetWorkflow';
 
-const IntroductionCard = defineAsyncComponent(
-  () => import('@/components/Cards/IntroductionCard.vue'),
-);
-const NotFoundCard = defineAsyncComponent(
-  () => import('@/components/Cards/NotFoundCard.vue'),
-);
-const NotificationCard = defineAsyncComponent(
-  () => import('@/components/Cards/NotificationCard.vue'),
-);
-const UserOrganizationInfo = defineAsyncComponent(
-  () => import('@/components/Cards/UserOrganizationInfo.vue'),
-);
+const IntroductionCard = defineAsyncComponent(() => import('@/components/Cards/IntroductionCard.vue'));
+const NotFoundCard = defineAsyncComponent(() => import('@/components/Cards/NotFoundCard.vue'));
+const NotificationCard = defineAsyncComponent(() => import('@/components/Cards/NotificationCard.vue'));
+const UserOrganizationInfo = defineAsyncComponent(() => import('@/components/Cards/UserOrganizationInfo.vue'));
+
+const workflowStore = useDatasetWorkflowStore();
 
 export default {
   name: 'DashboardPage',
@@ -375,11 +311,7 @@ export default {
   },
   computed: {
     ...mapState(['config']),
-    ...mapState(USER_SIGNIN_NAMESPACE, [
-      'user',
-      'userLoading',
-      'userEditLoading',
-    ]),
+    ...mapState(USER_SIGNIN_NAMESPACE, ['user', 'userLoading', 'userEditLoading']),
     ...mapState(USER_NAMESPACE, [
       'collaboratorDatasetIdsLoading',
       'collaboratorDatasetIds',
@@ -391,11 +323,10 @@ export default {
       'lastEditedDataset',
       'lastEditedDatasetPath',
     ]),
-    ...mapGetters(METADATA_NAMESPACE, [
-      'allTags',
-      'updatingTags',
-      'metadatasContent',
-    ]),
+    ...mapGetters(METADATA_NAMESPACE, ['allTags', 'updatingTags', 'metadatasContent']),
+    lastDataset() {
+      return this.lastEditedDataset;
+    },
     userDashboardConfig() {
       return this.config?.userDashboardConfig || {};
     },
@@ -407,6 +338,9 @@ export default {
     },
     datasetCreationActive() {
       return this.userEditMetadataConfig?.datasetCreationActive || false;
+    },
+    newWorkflowActive() {
+      return this.config.userEditMetadataConfig?.newWorkflowActive || false;
     },
     loading() {
       return (
@@ -424,10 +358,7 @@ export default {
 
       const errorDetail = `${this.userDatasetsError}<br /> <strong>Try reloading the datasets. If the problem persists please let use know via envidat@wsl.ch!</strong>`;
 
-      const notification = errorMessage(
-        'Error Loading Your Datasets',
-        errorDetail,
-      );
+      const notification = errorMessage('Error Loading Your Datasets', errorDetail);
       notification.timeout = 0;
 
       return notification;
@@ -513,10 +444,7 @@ export default {
         return null;
       }
 
-      return getUserOrganizationRoleMap(
-        this.user.id,
-        this.organizationsStore.userOrganizations,
-      );
+      return getUserOrganizationRoleMap(this.user.id, this.organizationsStore.userOrganizations);
     },
     organizationRoles() {
       if (!this.userOrganizationRoles) {
@@ -554,18 +482,13 @@ export default {
       return this.collaboratorDatasets?.length > 0;
     },
     isEditorAndAbove() {
-      return (
-        hasOrganizationRoles(this.organizationRoles) &&
-        !isMember(this.organizationRoles)
-      );
+      return hasOrganizationRoles(this.organizationRoles) && !isMember(this.organizationRoles);
     },
     noDatasetsInfos() {
       return {
         title: 'No Datasets',
         description: "It seems you don't have any datasets.",
-        actionDescription: this.isEditorAndAbove
-          ? 'Get started and create a new dataset'
-          : '',
+        actionDescription: this.isEditorAndAbove ? 'Get started and create a new dataset' : '',
         actionButtonText: 'New Dataset',
         image: UserNotFound2,
       };
@@ -580,9 +503,14 @@ export default {
       return this.user?.fullName?.split(' ')[1] || '';
     },
     hasLocalDataset() {
+      if (this.newWorkflowActive) {
+        const localDataset = this.workflowStore.localStorageService.dataset;
+        return !!localDataset;
+      }
+
       const localStorageData = getPreviewDatasetFromLocalStorage();
 
-      // eslint-disable-next-line no-unused-expressions
+      // eslint-disable-next-line @typescript-eslint/no-unused-expressions
       this.localDatasetUpdateCount;
 
       const properties = Object.keys(localStorageData);
@@ -590,14 +518,15 @@ export default {
       return properties.length > 0;
     },
     currentLocalDataset() {
-      const localStorageData = getPreviewDatasetFromLocalStorage();
+      let localDataset = this.newWorkflowActive
+        ? this.workflowStore.localStorageService.dataset
+        : getPreviewDatasetFromLocalStorage();
 
-      const localDataset = {
-        title: localStorageData[METADATA_TITLE_PROPERTY]
-          ? localStorageData[METADATA_TITLE_PROPERTY]
-          : '[Untitled Dataset]',
-        subTitle: localStorageData.description,
-        tags: localStorageData.keywords,
+      localDataset = {
+        ...localDataset,
+        title: localDataset.title ? localDataset.title : '[Untitled Dataset]',
+        subTitle: localDataset.description,
+        tags: localDataset.keywords,
         flatLayout: true,
       };
 
@@ -607,12 +536,7 @@ export default {
     },
   },
   methods: {
-    getPopularTagsFromDatasets(
-      datasets,
-      minCount = undefined,
-      maxCount = undefined,
-      maxTagAmount = 30,
-    ) {
+    getPopularTagsFromDatasets(datasets, minCount = undefined, maxCount = undefined, maxTagAmount = 30) {
       let tags = getPopularTags(datasets, '', minCount, maxCount);
 
       if (tags.length <= 0) {
@@ -620,10 +544,7 @@ export default {
       }
 
       if (tags.length > this.maxFilterTags) {
-        tags = tags.splice(
-          this.maxFilterTags,
-          tags.length - this.maxFilterTags,
-        );
+        tags = tags.splice(this.maxFilterTags, tags.length - this.maxFilterTags);
       }
 
       for (let j = 0; j < tags.length; j++) {
@@ -640,16 +561,10 @@ export default {
       return getMetadataVisibilityState(metadata);
     },
     contentFilteredByTags(value, selectedUserTagNames) {
-      return (
-        value.tags &&
-        tagsIncludedInSelectedTags(value.tags, selectedUserTagNames)
-      );
+      return value.tags && tagsIncludedInSelectedTags(value.tags, selectedUserTagNames);
     },
     loadRouteWithTags() {
-      const routeTags = loadRouteTags(
-        this.$route.query.tags,
-        this.selectedUserTagNames,
-      );
+      const routeTags = loadRouteTags(this.$route.query.tags, this.selectedUserTagNames);
 
       if (routeTags?.length > 0) {
         this.selectedUserTagNames = routeTags;
@@ -678,16 +593,10 @@ export default {
       });
 
       // always call the USER_GET_COLLABORATOR_DATASETS action because it resolves the store & state also when collaboratorDatasetIds is empty
-      await this.$store.dispatch(
-        `${USER_NAMESPACE}/${USER_GET_COLLABORATOR_DATASETS}`,
-        this.collaboratorDatasetIds,
-      );
+      await this.$store.dispatch(`${USER_NAMESPACE}/${USER_GET_COLLABORATOR_DATASETS}`, this.collaboratorDatasetIds);
     },
     async fetchUserOrganizationId(forceReload = false) {
-      if (
-        forceReload ||
-        (!forceReload && this.organizationsStore.userOrganizations?.length > 0)
-      ) {
+      if (forceReload || (!forceReload && this.organizationsStore.userOrganizations?.length > 0)) {
         await this.organizationsStore.UserGetOrgIds(this.user.id);
       }
     },
@@ -707,7 +616,12 @@ export default {
     },
     createClickCallback() {
       if (this.datasetCreationActive) {
-        this.$router.push({ path: METADATA_CREATION_PATH, query: '' });
+        const name = this.newWorkflowActive ? WORKFLOW_PAGENAME : METADATA_CREATION_PAGENAME;
+
+        this.$router.push({
+          name,
+          query: '',
+        });
       } else {
         window.open(`${this.ckanDomain}${this.createCKANUrl}`, '_blank');
       }
@@ -726,12 +640,21 @@ export default {
         // this.catchEditingClick(this.lastEditedDataset);
       }
     },
-    catchEditingClick(selectedDataset) {
+    catchEditingClick(selectedDatasetId) {
+      let name = METADATAEDIT_PAGENAME;
+      const params = {
+        metadataid: selectedDatasetId,
+      };
+
+      if (this.newWorkflowActive) {
+        name = WORKFLOW_PAGENAME;
+        params.id = selectedDatasetId;
+        delete params.metadataid;
+      }
+
       this.$router.push({
-        name: METADATAEDIT_PAGENAME,
-        params: {
-          metadataid: selectedDataset,
-        },
+        name,
+        params,
         query: {
           backPath: this.$route.fullPath,
         },
@@ -744,15 +667,10 @@ export default {
     },
 
     catchTagCloseClicked(tagName) {
-      this.selectedUserTagNames = this.selectedUserTagNames.filter(
-        (tag) => tag !== tagName,
-      );
+      this.selectedUserTagNames = this.selectedUserTagNames.filter((tag) => tag !== tagName);
     },
     catchMetadataClicked(datasetname) {
-      this.$store.commit(
-        `${METADATA_NAMESPACE}/${SET_DETAIL_PAGE_BACK_URL}`,
-        this.$route,
-      );
+      this.$store.commit(`${METADATA_NAMESPACE}/${SET_DETAIL_PAGE_BACK_URL}`, this.$route);
 
       this.$router.push({
         name: METADATADETAIL_PAGENAME,
@@ -770,25 +688,36 @@ export default {
           email: updateObject.data.email,
         };
 
-        this.$store.dispatch(
-          `${USER_SIGNIN_NAMESPACE}/${USER_EDITING_UPDATE}`,
-          payload,
-        );
+        this.$store.dispatch(`${USER_SIGNIN_NAMESPACE}/${USER_EDITING_UPDATE}`, payload);
       }
     },
     catchLocalCardClick() {
+      const name = this.newWorkflowActive ? WORKFLOW_PAGENAME : METADATA_CREATION_PAGENAME;
+      const params = {};
+
+      if (this.newWorkflowActive) {
+        params.id = this.currentLocalDataset.id;
+      }
+
       this.$router.push({
-        name: METADATA_CREATION_PAGENAME,
+        name,
+        params,
+        query: {
+          backPath: this.$route.fullPath,
+        },
       });
     },
     catchClearLocalStorage() {
       eventBus.emit(SHOW_DIALOG, {
         title: 'Clear your dataset in creation?',
-        message:
-          'This dataset is not stored on the server yet! Are you sure you want to delete it?',
+        message: 'This dataset is not stored on the server yet! Are you sure you want to delete it?',
         callback: () => {},
         cancelCallback: () => {
-          localStorage.clear();
+          if (this.newWorkflowActive) {
+            this.workflowStore.clearLocalStorage();
+          } else {
+            localStorage.clear();
+          }
 
           // increase this number to force the computed property to recalulate because the
           // localstorage is cleared. Localstorage isn't reactive so the computed prop
@@ -847,8 +776,7 @@ export default {
     },
     noCollaboratorDatasetsInfos: {
       title: 'No Collaborator Datasets',
-      description:
-        "It seems you don't have datasets where you are added as a collaborator.",
+      description: "It seems you don't have datasets where you are added as a collaborator.",
       image: UserNotFound2,
     },
     noOrganizationsInfos: {
@@ -865,6 +793,7 @@ export default {
     ],
     localDatasetUpdateCount: 0,
     mdiRefresh,
+    workflowStore,
   }),
   components: {
     MetadataList,
